requirements.txt
# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Database
sqlmodel>=0.0.16

# Configuration
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Async Tasks
celery[redis]>=5.3.0
redis>=5.0.0

# Video Processing
moviepy>=2.0.0
Pillow>=10.0.0

# LLM & AI
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-google-genai>=1.0.0
openai>=1.12.0
google-generativeai>=0.4.0
google-genai>=1.0.0

# HTTP Client
requests>=2.31.0
httpx>=0.26.0
aiohttp>=3.9.0

# Audio Processing
pydub>=0.25.1

# Utilities
tenacity>=8.2.0
structlog>=24.1.0

# Development
ruff>=0.2.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
